requirements.txt
torch
torchvision
numpy
opencv-python
pillow
matplotlib
tensorflow
keras